Formatting

+++ src/plugin/index.ts
@@ -362,13 +362,13 @@ export function configureC8yScreenshotPlugin(
           }
           logScreenshot(`Deleted ${source}`);
         });
-      }
+      };
       // path contains spec name, remove it. might only be required for run() mode however
       const newPath =
         details.specName.trim() == ""
           ? details.path
           : details.path?.replace(`${details.specName}${path.sep}`, "");
-          logRun(`details.path: ${details.path} -> newPath: ${newPath}`);
+      logRun(`details.path: ${details.path} -> newPath: ${newPath}`);
 
       const screenshotTarget = path.dirname(newPath);
       const diffTarget =
@@ -431,8 +431,7 @@ export function configureC8yScreenshotPlugin(
                 `Skipping ${screenshotFile} (skipMove: ${diffOptions.skipMove})`
               );
               // discard the screenshot if there is a match and skipMove is false
-              if(!diffOptions.skipMove)
-                deleteFile(details.path);
+              if (!diffOptions.skipMove) deleteFile(details.path);
             } else {
               moveFile(details.path, screenshotFile);
             }
